feat: add NSID strict parser with authority and name accessors
